# app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FX Aggregator",
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost"
        }
    ],
    "paths": {
        "/api/convert/{base}/to/{quote}": {
            "post": {
                "tags": [
                    "convertor"
                ],
                "summary": "Submit conversion request",
                "operationId": "convert",
                "parameters": [
                    {
                        "name": "base",
                        "in": "path",
                        "description": "Currency to convert FROM",
                        "required": true,
                        "example": "EUR"
                    },
                    {
                        "name": "quote",
                        "in": "path",
                        "description": "Currency to convert TO",
                        "required": true,
                        "example": "RON"
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "example": "2022-01-03"
                    }
                ],
                "requestBody": {
                    "description": "Conversion amount is required",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "value"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2022-01-03"
                                        },
                                        "value": {
                                            "type": "number",
                                            "format": "float",
                                            "example": "4.94"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/currencies": {
            "get": {
                "tags": [
                    "currencies"
                ],
                "summary": "List of supported currencies",
                "operationId": "currencies",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "RON"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/currencies/{currency}/rate": {
            "get": {
                "tags": [
                    "currencies"
                ],
                "summary": "Retrieve the rate again RON at a given date",
                "operationId": "rate",
                "parameters": [
                    {
                        "name": "currency",
                        "in": "path",
                        "description": "Currency for which to see the RON value",
                        "required": true,
                        "example": "EUR"
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "required": true,
                        "example": "2022-01-03"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2022-01-03"
                                        },
                                        "value": {
                                            "type": "number",
                                            "format": "float",
                                            "example": "4.94"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}